requirements.txt
pandas
mutagen
python-dotenv
librosa
demucs
soundfile
tqdm
numpy
jupyter
ipykernel